test(tbab_cation.xyz): add tetrabutylammonium cation
During the setup of Pytest checks, the structure described by
the Hill formula C9H11 was recognized as ambigious.  With the
tools at disposition, it could present either an isopropylbenzene
cation, or an isopropylbenzene anion.  With CID 140141, PubChem
indexes the neutral 2-phenyl-2-propyl radical, too.  This is why
the test set of structres was edited to include the example of
tetrabutylammonium cation instead.

Signed-off-by: Norwid Behrnd <nbehrnd@yahoo.com>