test: replace checks on [NBu4]^+ by [NMe4]^+ cation
To keep the checks compact, checks on the recovery of the
tetrabutylammonium cation are replaced by ones to probe the
tetramethylammonium cation instead.  The exchange doesn't
matter for the checks in terms of the conversion, or chemistry
except that checks may be more compact for managing only about
a third of the atoms per molecule.

Signed-off-by: Norwid Behrnd <nbehrnd@yahoo.com>